requirements.txt
python-pptx
readability-lxml
BeautifulSoup
markdownify
azure.cognitiveservices.speech
AzureOpenAI
python-dotenv
openai